add FuturesIoCompatExt trait
