+++ src/io/futures.rs
@@ -1,11 +1,50 @@
-use super::{AsyncSocket, Compat};
+use super::AsyncSocket;
 use futures_io::{AsyncRead, AsyncWrite};
 use std::{
     io::Result as IoResult,
-    ops::DerefMut,
+    ops::{Deref, DerefMut},
     pin::Pin,
     task::{Context, Poll},
 };
+
+pub struct Compat<S>(S);
+
+impl<S> Compat<S> {
+    pub fn new(inner: S) -> Self {
+        Compat(inner)
+    }
+
+    pub fn into_inner(self) -> S {
+        self.0
+    }
+}
+
+impl<S> Deref for Compat<S> {
+    type Target = S;
+
+    fn deref(&self) -> &Self::Target {
+        &self.0
+    }
+}
+
+impl<S> DerefMut for Compat<S> {
+    fn deref_mut(&mut self) -> &mut Self::Target {
+        &mut self.0
+    }
+}
+
+pub trait FuturesIoCompatExt {
+    fn compat(self) -> Compat<Self>
+    where Self: Sized;
+}
+
+impl<S> FuturesIoCompatExt for S
+where S: AsyncRead + AsyncWrite + Unpin
+{
+    fn compat(self) -> Compat<Self> {
+        Compat::new(self)
+    }
+}
 
 impl<S> AsyncSocket for Compat<S>
 where S: AsyncRead + AsyncWrite + Unpin
