+++ src/io/mod.rs
@@ -8,36 +8,12 @@ use std::{
     future::Future,
     io::{Error, ErrorKind},
     mem,
-    ops::{Deref, DerefMut},
     pin::Pin,
     task::{Context, Poll},
 };
 
-pub struct Compat<S>(S);
-
-impl<S> Compat<S> {
-    pub fn new(inner: S) -> Self {
-        Compat(inner)
-    }
-
-    pub fn into_inner(self) -> S {
-        self.0
-    }
-}
-
-impl<S> Deref for Compat<S> {
-    type Target = S;
-
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-impl<S> DerefMut for Compat<S> {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        &mut self.0
-    }
-}
+#[cfg(feature = "futures-io")]
+pub use futures::{Compat, FuturesIoCompatExt};
 
 pub trait AsyncSocket {
     fn poll_read(self: Pin<&mut Self>, cx: &mut Context<'_>, buf: &mut [u8]) -> Poll<Result<usize, Error>>;
