+++ tests/no_auth.rs
@@ -47,9 +47,8 @@ fn bind_with_socket_no_auth() -> Result<()> {
 #[cfg(feature = "futures-io")]
 #[test]
 fn connect_with_socket_no_auth_futures_io() -> Result<()> {
-    use tokio_socks::io::Compat;
     let runtime = futures_utils::runtime().lock().unwrap();
-    let socket = Compat::new(runtime.block_on(futures_utils::connect_unix(UNIX_PROXY_ADDR))?);
+    let socket = runtime.block_on(futures_utils::connect_unix(UNIX_PROXY_ADDR))?.compat();
     let conn = runtime.block_on(Socks5Stream::connect_with_socket(socket, ECHO_SERVER_ADDR))?;
     runtime.block_on(futures_utils::test_connect(conn))
 }
@@ -57,10 +56,9 @@ fn connect_with_socket_no_auth_futures_io() -> Result<()> {
 #[cfg(feature = "futures-io")]
 #[test]
 fn bind_with_socket_no_auth_futures_io() -> Result<()> {
-    use tokio_socks::io::Compat;
     let bind = {
         let runtime = futures_utils::runtime().lock().unwrap();
-        let socket = Compat::new(runtime.block_on(futures_utils::connect_unix(UNIX_PROXY_ADDR))?);
+        let socket = runtime.block_on(futures_utils::connect_unix(UNIX_PROXY_ADDR))?.compat();
         runtime.block_on(Socks5Listener::bind_with_socket(socket, ECHO_SERVER_ADDR))
     }?;
     futures_utils::test_bind(bind)
