+++ tests/socks4_userid.rs
@@ -60,9 +60,10 @@ fn bind_with_socket_userid() -> Result<()> {
 #[cfg(feature = "futures-io")]
 #[test]
 fn connect_with_socket_userid_futures_io() -> Result<()> {
-    use tokio_socks::io::Compat;
     let runtime = futures_utils::runtime().lock().unwrap();
-    let socket = Compat::new(runtime.block_on(futures_utils::connect_unix(UNIX_SOCKS4_PROXY_ADDR))?);
+    let socket = runtime
+        .block_on(futures_utils::connect_unix(UNIX_SOCKS4_PROXY_ADDR))?
+        .compat();
     let conn = runtime.block_on(Socks4Stream::connect_with_userid_and_socket(
         socket,
         ECHO_SERVER_ADDR,
@@ -74,10 +75,11 @@ fn connect_with_socket_userid_futures_io() -> Result<()> {
 #[cfg(feature = "futures-io")]
 #[test]
 fn bind_with_socket_userid_futures_io() -> Result<()> {
-    use tokio_socks::io::Compat;
     let bind = {
         let runtime = futures_utils::runtime().lock().unwrap();
-        let socket = Compat::new(runtime.block_on(futures_utils::connect_unix(UNIX_SOCKS4_PROXY_ADDR))?);
+        let socket = runtime
+            .block_on(futures_utils::connect_unix(UNIX_SOCKS4_PROXY_ADDR))?
+            .compat();
         runtime.block_on(Socks4Listener::bind_with_user_and_socket(
             socket,
             ECHO_SERVER_ADDR,
